feat: exit tab mode if we're down to a single tab
This is a UX improvement where we only show tabs if there is more than one. There's no reason to have a single tab, since we have just one dashboard.

When exiting tab mode from having one tab, we keep that tab in the database. For now that means a user cannot delete the final tab so that we keep the tab name. Maybe this isn't all that important and we can revist later.