feat: Introduce ai agent user project preferences table (#15402)
### Description:
This PR adds user preferences for AI agents. This will be used to allow users to set and retrieve their default agent for a project. I decided to introduce a specialized agent user preferences table as this is an EE feature, and didn't want to expose that configuration table to other users/customers

- New database table `ai_agent_user_preferences` to store user preferences
- API endpoints to get and set the default agent:
  - GET `/api/v1/projects/:projectUuid/aiAgents/preferences`
  - POST `/api/v1/projects/:projectUuid/aiAgents/preferences` with `{ defaultAgentUuid: string }` body

I'll use these methods in a follow up PR to update the current logic relying on LocalStorage (to fix #15386)