feat: allow create projects without warehouse credentials (#15594)
Working towards https://github.com/lightdash/lightdash/issues/15347

This makes it possible to create a project without any warehouse credentials. This is only available on the "precompiled" endpoint, which expects a bunch of already constructed explorers.

The `Project` variable already had the warehouse credentials as optional and it appears to be well handled.

Changes:
- LightdashAnalytics - now warehouse type is optional so will be missing sometimes
- `OrganizationProject` had warehouse credentials as required but I've made it optional (only needed for user specific credentials flow in the UI)
- Added defaults to the `ProjectConnection` form in case there are no credentials.