feat: support `tags` + `meta` under `config` (#15291)
I compiled dbt projects with different `dbt` versions. The table below summarises the results.

The `yaml source key` is the location of the config in the yaml files. I've included the "old way" using `meta` and `tags` and the "new way" where those keys are nested under `config`. The other columns show the location of the properties in the `manifest.json` - so it shows how the configuration is projected from the yaml files to the `manifest.json` (which is what we use as the source of truth for dbt metadata).

| yaml source key    | dbt 1.9.8 & dbt 1.10.0-rc2                 | dbt fusion          |
|--------------------|-------------------------------------|---------------------|
| model.meta (old)   | .model.meta ⚠️<br>.model.config.meta | ❌                   |
| model.tags (old)   | ❌                                   | ❌                   |
| column.meta (old)  | .column.meta                        | ❌                   |
| column.tags (old)  | .column.tags                        | ❌                   |
| model.config.meta  | .model.meta ⚠️<br>.model.config.meta | ❌⚠️                  |
| model.config.tags  | .model.tags <br>.model.config.tags  | .model.config.tags  |
| column.config.meta | .column.config.meta                 | .column.config.meta |
| column.config.tags | .column.config.tags                 | .column.config.tags |

⚠️ **Cannot** use old and new style meta syntax together. In dbt 1.9 both old and new style are sent to .model.meta and .model.config.meta
❌⚠️ **BUG** in dbt fusion where model meta doesn’t flow through

To maintain backwards compatibility:

model config is read as `.model.config.meta || .model.meta` 
model tags are read as `.model.config.tags || .model.tags`
column meta is read as `.column.config.meta || .column.meta`
column tags are read as `.column.config.tags || .column.tags`

This maintains backwards compatibility for older dbt versions while preferring the new syntax

Steps for users migrating:
- Move `meta` and `tags` for your models and columns to `config.meta` and `config.tags` 
- Both old and new style are supported in dbt 1.9 and dbt 1.10 - so you can migrate incrementally
- Just don’t use old + new style on the same model/column as this will prevent compiling

I'll update docs separately with a migration guide

test-cli test-backend test-frontend