feat: Link saved charts to ai prompts (#15609)
<!-- Thanks so much for your PR, your contribution is appreciated! ❤️ -->

Closes: #15552, #15452

### Description:

Added the ability to associate saved charts with AI agent messages. When a user saves a chart from an AI agent response, the saved chart UUID is now stored in the database and linked to the message. 
This way, users can see which charts have been saved and provides a direct link to view the saved chart.

#### TODO
- [ ] Refresh message on saved chart, so we instantly switch the `save chart` with the `view chart` button
- [ ] Thoroughly testing