fix: pass maxQueryLimit to getExploreBlocks and parse vizConfig properly (#15751)
### Description:
Fixed the Slack explore blocks functionality by passing the maximum query limit to the `getExploreBlocks` function. This ensures that the visualization configuration is properly parsed with the correct query limits when generating explore links in Slack responses.

The change also updates the parsing of the visualization configuration to use the `parseVizConfig` utility instead of directly casting to `MetricQuery`, providing better type safety and validation.

before

![CleanShot 2025-07-07 at 18.43.49@2x.png](https://graphite-user-uploaded-assets-prod.s3.amazonaws.com/cOL6QbQzyaUdOs5psinD/e791e157-5de8-44a7-8038-2396721b9b2a.png)

after

![CleanShot 2025-07-07 at 18.43.44@2x.png](https://graphite-user-uploaded-assets-prod.s3.amazonaws.com/cOL6QbQzyaUdOs5psinD/6e236410-6260-40c9-a753-4d0c97e659a9.png)