fix: Allow empty value for BigQuery maximunBytesBilled (#16164)
### Description:
Makes the `maximumBytesBilled` field optional in the BigQuery form by removing the `required` attribute. Updates the help text to clarify that leaving the field empty or with a value of 0 means no limit will be applied. Also simplifies the logic in the BigqueryWarehouseClient to handle empty or zero values properly.