feat: avoid layout jump when loading new results using ai search box (#17441)
<!-- Thanks so much for your PR, your contribution is appreciated! ❤️ -->

### Description

Improved search UX when using the AI Search box by using the `keepPreviousData` flag when doing requests. Instead of cleaning the list of results while fetching, we now keep it and display a loader in the input to avoid layout jumps.

_Before_
[Screen Recording 2025-10-14 at 14.55.54.mov](https://app.graphite.dev/user-attachments/video/07ab6f7f-726e-48da-aff9-058b471b12dc.mov)

_After_
[Screen Recording 2025-10-14 at 14.55.06.mov](https://app.graphite.dev/user-attachments/video/54ca92cf-ca53-4a1e-832f-2ed826530946.mov)

(Extra latency was added into the requests to better display the loading state)