feat: see compilation raw information (#17586)
### Description:
Added a compilation log details drawer to the Compilation History Table. Users can now click on any row in the table to view detailed information about the compilation log in a side drawer. The drawer includes a copy button to easily copy the JSON data and uses ReactJson for a collapsible, formatted display of the log details.

![image.png](https://app.graphite.dev/user-attachments/assets/72e514f7-9110-4070-93a8-f3e8b40c1c0e.png)