feat: add charts to embedding table (#17596)
<!-- Thanks so much for your PR, your contribution is appreciated! ❤️ -->

Closes: CENG-99<!-- reference the related issue e.g. #150 -->

### Description: 

This is the first PR in making embed charts available in Lightdash. 

The migration adds two new columns to the embedding table to allow configuring a chart for embedding. While this isn't a great long-term solution (adding more `*_uuids`​ fields), this is a bridge solution until we have a stronger long-term solution for sharing content. 

<!-- Add a description of the changes proposed in the pull request. -->

<!-- Even better add a screenshot / gif / loom -->