fix: set dashboard tabs in embeds (#17720)
<!-- Thanks so much for your PR, your contribution is appreciated! ❤️ -->

Closes: <!-- reference the related issue e.g. #150 -->

### Description:
<!-- Add a description of the changes proposed in the pull request. -->

<!-- Even better add a screenshot / gif / loom -->

At some point dashboard tabs were not being set in DashboardProvider. This created a bug where we no longer loaded an embedded tab in the iframe if provided—we just defer to the first tab. 

This explicitly sets dashboard tabs if they're empty and have sorted tabs in the EmbedDashboard.