fix: remove jwt from URL hash (#17652)
<!-- Thanks so much for your PR, your contribution is appreciated! ❤️ -->

Closes: CENG-116<!-- reference the related issue e.g. #150 -->

### Description:

<!-- Add a description of the changes proposed in the pull request. -->

<!-- Even better add a screenshot / gif / loom -->

We had been relying on implicit behavior in DashboardProvider to remove the JWT from the URL hash. Something changed recently and we no longer implicitly make that redirect.

This PR explicitly removes the JWT from the URL. We use the browser directly rather than react-router since we're already using window.location to get the URL hash.