fix: random black line appears on top of some charts (#17943)
<!-- Thanks so much for your PR, your contribution is appreciated! ❤️ -->

Closes: #17666

### Description:
Fixed an issue where a random line appears in charts (secondary axis). This occurs when configuring a chart with min/max axis values. When the library detects a min/max configuration, it renders the secondary axis using default colors. This PR improves secondary axis rendering to prevent this behavior while accounting for cases where the axis should be flipped.

_after_
![image.png](https://app.graphite.com/user-attachments/assets/d37c4881-2f29-46d3-b20c-e4258ba93a4f.png)

_before_

![image.png](https://app.graphite.com/user-attachments/assets/04a96efc-8538-4696-92fe-8d9600cd8639.png)