feat: add Chart to embed SDK (#18017)
<!-- Thanks so much for your PR, your contribution is appreciated! ❤️ -->

Closes: CENG-104

### Description:

<!-- Add a description of the changes proposed in the pull request. -->

<!-- Even better add a screenshot / gif / loom -->

Adds a Chart component to the Lightdash Embed SDK. This is a two part PR with stricter permissions coming in a [follow up](https://linear.app/lightdash/issue/CENG-148/add-stricter-permissions-for-chart-embeds).

### Testing

I added an http-client env file. This works for both VSCode and WebStorm HTTP clients. You can create an http-client.env.private.json file to override values in the public file. 

Create an embed URL for the charts you need and use that in the SDK test app.

### Examples

#### Render Embedded Chart

JWTs for charts work only for charts. Dashboards will not load. 

![Screen Cast 2025-11-11 at 12.36.24 PM.gif](https://app.graphite.com/user-attachments/assets/c021e055-f280-45d5-9170-c86833955e5c.gif)

#### Allow chart slugs in addition to UUIDs

Allow both UUIDs and slugs. Debugging breakpoints are optional. 

![Screen Cast 2025-11-11 at 12.37.11 PM.gif](https://app.graphite.com/user-attachments/assets/118c40a1-2772-4c4c-a3e1-d0d9c2b22856.gif)

#### Only allow chart UUID in JWT

Only load embedded charts for a chart UUID or slug that matches what's bound to the JWT

![Screen Cast 2025-11-11 at 12.37.58 PM.gif](https://app.graphite.com/user-attachments/assets/150323c4-01f4-4254-b509-8447b9cafca1.gif)