fix: custom metric sql templates and filters not updated on rename (#18146)
### Description:
when renaming a field, custom metric sql templates and filter `fieldRef` values were not being updated if the db project was deployed before running rename command. 
We were constructing the `fromReference` using the field name from the explore metadata. When the project was updated before running rename command, the explore already contained the new field name. This caused `fromReference` to be set to the new name instead of the old name, preventing the string matching from finding the old field referenced in charts