chore: parallelize docker PR builds (#18201)
<!-- Thanks so much for your PR, your contribution is appreciated! ❤️ -->

Closes: <!-- reference the related issue e.g. #150 -->

### Description:

<!-- Add a description of the changes proposed in the pull request. -->

<!-- Even better add a screenshot / gif / loom -->

Optimize PR builds to run in parallel stages where we can. If this is successful, we can roll it up to the main dockerfile—and even consolidate.

It looks like the mount caches added in this PR are successfully bring down build times on cache hits. We should monitor build times to see if further optimizations are neccessary or not. 

test-frontend test-backend