chore: make project create:Project distinctly for previews (#18676)
<!-- Thanks so much for your PR, your contribution is appreciated! ❤️ -->

Closes: #18540
Closes: [https://linear.app/lightdash/issue/CENG-240/add-scope-for-preview-projects](https://linear.app/lightdash/issue/CENG-240/add-scope-for-preview-projects)

### Description:

<!-- Add a description of the changes proposed in the pull request. -->

<!-- Even better add a screenshot / gif / loom -->

We already supported project members below Developer only getting `create:Project` for previews. But it wasn't obvious because it was simply `create:Project`. This makes the scope transparent with `create:Project@preview`. We remove the modifier when apply CASL abilities, so we get the same behavior. Now it's just documented and clear.

The giant files changed is just whitespace. I removed a `describe` from the scope builder test. This change isn't that big :sweat_smile:

## Testing

1. Duplicate the Editor role as a custom role and add `create:Project@preview`
2. Assign the custom role to an Editor user
3. Login via the CLI 
4. `lightdash deploy --create foo` should fail
5. `lightdash preview` should pass