feat: add styles to sticky dashboard elements (#19027)
### Description:

Added a new `StickyWithDetection` component that detects when a sticky element becomes stuck to improve the dashboard tabs UI. This component uses the Intersection Observer API to detect when an element becomes sticky and provides this state via a render prop.

Implemented this component in the dashboard tabs to add a bottom border when the tabs are stuck to the top of the viewport, improving visual distinction when scrolling. The component handles different offset values based on whether the dashboard is in edit mode.

#### With tabs

[CleanShot 2025-12-23 at 11.28.52.mp4 <span class="graphite__hidden">(uploaded via Graphite)</span> <img class="graphite__hidden" src="https://app.graphite.com/user-attachments/thumbnails/48d6e033-62bb-4fb0-b3be-1ba7ad461f85.mp4" />](https://app.graphite.com/user-attachments/video/48d6e033-62bb-4fb0-b3be-1ba7ad461f85.mp4)

#### No Tabs

[CleanShot 2025-12-23 at 11.30.16.mp4 <span class="graphite__hidden">(uploaded via Graphite)</span> <img class="graphite__hidden" src="https://app.graphite.com/user-attachments/thumbnails/c054920c-c7d5-4b98-ac4e-b3026793e805.mp4" />](https://app.graphite.com/user-attachments/video/c054920c-c7d5-4b98-ac4e-b3026793e805.mp4)